Added config.php file - not a webpage but rather a container for database-related credentials and configurations.
Note that each individual will likely encounter issues with database connectivity - this file stores the server address, username and password used to connect to the database server, which will vary between team members. As such for testing kindly remember to store your own credentials and locations in the file and rotate between everyone's appropriately, or use a centralised username and password.